refactor: extract formatReportDate to shared lib/formatters
Eliminates change amplification - single source of truth for date
formatting instead of duplicated code in page and widget.
