+++ lib/formatters.ts
@@ -1,0 +1,16 @@
+export function formatReportDate(startDate: number, endDate: number): string {
+  const start = new Date(startDate);
+  const end = new Date(endDate);
+
+  const sameDay = start.toDateString() === end.toDateString();
+
+  if (sameDay) {
+    return end.toLocaleDateString("en-US", {
+      month: "long",
+      day: "numeric",
+      year: "numeric",
+    });
+  }
+
+  return `${start.toLocaleDateString("en-US", { month: "short", day: "numeric" })} – ${end.toLocaleDateString("en-US", { month: "short", day: "numeric", year: "numeric" })}`;
+}
